+++ doc/changelog.rst
@@ -3,6 +3,7 @@
 Release Notes
 =============
 
+- :support:`-` pretalx will no longer make alarmist noises about missing migrations.
 - :feature:`882` You can now filter the submission list by specific question answers.
 - :feature:`-` Questions that are visible to reviewers now also show up on the review page, to avoid having to switch to the full submission view in the review process.
 - :bug:`928` Reviewers had access to question answers that were not explicitly visible to reviewers by way of the API.